New Index File
This is the new homepage. It's basically a copy of tree.php, but it automatically will show ID 4's tree as an example tree. I added a message at the top that this is an example tree so people can get a feel for what the site does immediatly upon loading the page. 

As we make design changes to tree.php, we'll need to make sure index.php will look the same before launch.