Fixed the multi-click delete bug, new bug
Bug now: If you load the tree, delete button pops up great. If you scroll to the right, it breaks. Even if you refresh or scroll back, it stays broken. You HAVE to go to Edit Account and then Back to Tree for the delete button to work again, until you scroll to the right.
This is the weirdest bug by far.